added comments for J2ObjC added/removed lines, changed ZipUtil file into com.google.j2objc.util
