+++ jre_emul/android/platform/libcore/luni/src/test/java/libcore/java/util/zip/GZIPOutputStreamTest.java
@@ -23,14 +23,16 @@ import java.io.OutputStream;
 import java.util.Arrays;
 import java.util.Random;
 import java.util.zip.GZIPOutputStream;
-//import libcore.junit.junit3.TestCaseWithRules;
-//import libcore.junit.util.ResourceLeakageDetector;
+/* J2ObjC removed: not supported by Junit 4.11 (https://github.com/google/j2objc/issues/1318).
+import libcore.junit.junit3.TestCaseWithRules;
+import libcore.junit.util.ResourceLeakageDetector; */
 import org.junit.Rule;
 import org.junit.rules.TestRule;
 
-public final class GZIPOutputStreamTest extends junit.framework.TestCase /* TestCaseWithRules */ {
-//  @Rule
-//  public TestRule resourceLeakageDetectorRule = ResourceLeakageDetector.getRule();
+public final class GZIPOutputStreamTest extends junit.framework.TestCase /* J2ObjC removed: TestCaseWithRules */ {
+  /* J2ObjC removed: not supported by Junit 4.11 (https://github.com/google/j2objc/issues/1318).
+  @Rule
+  public TestRule resourceLeakageDetectorRule = ResourceLeakageDetector.getRule(); */
 
   public void testShortMessage() throws IOException {
     byte[] data = gzip(("Hello World").getBytes("UTF-8"));
