+++ jre_emul/android/platform/libcore/luni/src/test/java/libcore/java/util/zip/InflaterTest.java
@@ -20,14 +20,16 @@ import java.io.ByteArrayOutputStream;
 import java.util.zip.Adler32;
 import java.util.zip.Deflater;
 import java.util.zip.Inflater;
-//import libcore.junit.junit3.TestCaseWithRules;
-//import libcore.junit.util.ResourceLeakageDetector;
+/* J2ObjC removed: not supported by Junit 4.11 (https://github.com/google/j2objc/issues/1318).
+import libcore.junit.junit3.TestCaseWithRules;
+import libcore.junit.util.ResourceLeakageDetector; */
 import org.junit.Rule;
 import org.junit.rules.TestRule;
 
-public class InflaterTest extends junit.framework.TestCase /* TestCaseWithRules */ {
-//    @Rule
-//    public TestRule resourceLeakageDetectorRule = ResourceLeakageDetector.getRule();
+public class InflaterTest extends junit.framework.TestCase /* J2ObjC removed: TestCaseWithRules */ {
+    /* J2ObjC removed: not supported by Junit 4.11 (https://github.com/google/j2objc/issues/1318).
+    @Rule
+    public TestRule resourceLeakageDetectorRule = ResourceLeakageDetector.getRule(); */
 
     public void testDefaultDictionary() throws Exception {
         assertRoundTrip(null);
