+++ jre_emul/android/platform/libcore/ojluni/src/main/java/java/util/zip/ZipInputStream.java
@@ -33,7 +33,7 @@ import java.io.PushbackInputStream;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import static java.util.zip.ZipConstants64.*;
-import static java.util.zip.ZipUtils.*;
+import static com.google.j2objc.util.ZipUtils.*;
 
 /**
  * This class implements an input stream filter for reading files in the
