+++ jre_emul/jre_sources.mk
@@ -1676,11 +1676,11 @@ JAVA_PUBLIC_SOURCES_ZIP = \
   java/util/zip/ZipOutputStream.java
 
 JAVA_PRIVATE_SOURCES_ZIP = \
+  com/google/j2objc/util/ZipUtils.java \
   java/util/jar/JarVerifier.java \
   java/util/zip/ZStreamRef.java \
   java/util/zip/ZipCoder.java \
   java/util/zip/ZipConstants64.java \
-  java/util/zip/ZipUtils.java \
   libcore/io/Base64.java \
   libcore/io/BufferIterator.java \
   libcore/net/url/JarHandler.java \
